Move DatabaseTableName into its own file
Since it's no longer really a Label at all
