+++ soql-analyzer/src/main/scala/com/socrata/soql/analyzer2/Labels.scala
@@ -107,25 +107,6 @@ object AutoTableLabel {
   }
 }
 
-final case class DatabaseTableName[+T](name: T) {
-  def debugDoc(implicit ev: HasDoc[T]): Doc[Nothing] = ev.docOf(name)
-}
-
-object DatabaseTableName {
-  implicit def jEncode[T: JsonEncode] = WrapperJsonEncode[DatabaseTableName[T]](_.name)
-  implicit def jDecode[T: JsonDecode] = WrapperJsonDecode[DatabaseTableName[T]](DatabaseTableName[T](_))
-
-  implicit def serialize[T: Writable] = new Writable[DatabaseTableName[T]] {
-    def writeTo(buffer: WriteBuffer, d: DatabaseTableName[T]) =
-      buffer.write(d.name)
-  }
-
-  implicit def deserialize[T: Readable] = new Readable[DatabaseTableName[T]] {
-    def readFrom(buffer: ReadBuffer) =
-      DatabaseTableName(buffer.read[T]())
-  }
-}
-
 sealed abstract class ColumnLabel[+T] {
   def debugDoc(implicit ev: HasDoc[T]): Doc[Nothing]
 }
